requirements.txt
requests==2.31.0
httpie==3.2.2
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
streamlit==1.29.0
python-dotenv==1.0.0
schedule==1.2.0 